Some improvements done.
