+++ src/main/java/org/fundacionjala/coding/abel/movies/Children.java
@@ -1,0 +1,32 @@
+package org.fundacionjala.coding.abel.movies;
+
+/**
+ * Created by AbelBarrientos on 5/31/2017.
+ */
+class Children extends Movie {
+
+    public static final double DEFAULT_INITIAL_AMOUNT = 1.5;
+    public static final int MAX_DAYS = 3;
+
+    /**
+     * This is the default constructor.
+     * @param title String.
+     */
+    Children(String title) {
+        super(title);
+    }
+
+    /**
+     * This method calculates the amount for Children movie type.
+     * @param daysRented Number of days rented.
+     * @return Amount.
+     */
+    @Override
+    public double calculateAmount(int daysRented) {
+        thisAmount += DEFAULT_INITIAL_AMOUNT;
+        if (daysRented > MAX_DAYS) {
+            thisAmount += (daysRented - MAX_DAYS) * DEFAULT_INITIAL_AMOUNT;
+        }
+        return thisAmount;
+    }
+}
